% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_connection.R
\name{close_connection}
\alias{close_connection}
\title{Disconnects a given database connection.}
\usage{
close_connection(con)
}
\arguments{
\item{con}{A database connection (an object of class MySQLConnection).}
}
\value{
A logical vector of length 1, indicating success or failure.
}
\description{
Disconnects a given database connection.
}
\examples{
# establishing connection to oefenweb database
con <- connect(db = "oefenwebDatabase_test")
# terminating connnection to oefenweb database
close_connection(con)
}

