% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_connections.R
\name{close_connections}
\alias{close_connections}
\title{Disconnects all database connections.}
\usage{
close_connections()
}
\value{
A logical list, indicating success or failure for all connections.
}
\description{
Disconnects all database connections.
}
