% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connects to a given database.}
\usage{
connect(
  dbname = "oefenweb_nl_app",
  default.file = path.expand(file.path("~", ".my.cnf"))
)
}
\arguments{
\item{dbname}{A database name, for example "oefenweb_nl_app" or "mathsgarden_com_app".}

\item{default.file}{A MySQL configuration file, containing login credentials.}
}
\value{
A database connection (an object of class MySQLConnection).
}
\description{
Connects to a given database.
}
