% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Retrieve results from a query.}
\usage{
get_query(..., con, params = NULL)
}
\arguments{
\item{...}{[expressions] Unnamed arguments are taken to be expression string(s) to format.
Multiple inputs are concatenated together before formatting.}

\item{con}{A DBIConnection object, as returned by dbConnect().}

\item{params}{[list] Containing parameters called in SQL statement with \{params[[i]]\} or \{params[[j]]*\}}
}
\value{
[data.frame] With as many rows as records were fetched and as many columns as fields in the result set.
}
\description{
Suppresses warnings and uses glue_sql for safety.
}
\examples{
\dontrun{
domain_ids <- 1:4
game_name <- "multiplication"
get_query("SELECT `name` FROM `domains` WHERE `id` IN ({params[[1]]*}) AND `name` != {params[[2]]}",
          params = list(domain_ids, game_name))
}
}
